requirements.txt
flask==3.0.0
streamlit==1.29.0
openai==1.3.0
moviepy==1.0.3
langchain==0.1.0
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.0
